+++ packages/humans/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @abloatai/humans
 
+## 0.41.0
+
+### Patch Changes
+
+- @abloatai/transaction@0.41.0
+
 ## 0.40.0
 
 ### Patch Changes